Addresses Linus's comments
